requirements.txt
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
psycopg2-binary==2.9.9
sqlparse==0.4.4
